refactor module Core
