+++ compiler/Plato/Core/Closure.hs
@@ -5,8 +5,17 @@ import Data.Maybe
 import GHC.Stack
 
 import Plato.Common.Ident
-import Plato.Core.Data
+import Plato.Common.Pretty
 import Plato.Syntax.Core
+
+type Index = Int
+
+newtype Scope = Scope (IdentMap (Index, Maybe (Clos Type))) deriving (Eq, Show)
+
+type Clos a = (a, Scope)
+
+instance PrettyWithContext a => PrettyWithContext (Clos a) where
+        pretty' c (t, _) = pretty' c t
 
 emptyScope :: Scope
 emptyScope = Scope M.empty
@@ -35,7 +44,3 @@ instance Closure (Clos a) where
 instance Closure a => Closure (Bind a) where
         getScope (_, a) = getScope a
         putScope (x, a) s = (x, putScope a s)
-
-instance Closure Boxed where
-        getScope (Boxed c) = getScope c
-        putScope (Boxed c) = Boxed . putScope c
