+++ compiler/Plato/Typing/Tc.hs
@@ -135,8 +135,7 @@ tcRho ::
         LExpr 'Untyped ->
         Expected Rho ->
         m (LExpr 'Typed)
-tcRho (L sp exp) exp_ty = do
-        L sp <$> tcRho' exp exp_ty
+tcRho (L sp exp) exp_ty = L sp <$> tcRho' exp exp_ty
     where
         tcRho' :: Expr 'Untyped -> Expected Rho -> m (Expr 'Typed)
         tcRho' (VarE var) exp_ty = do
